// finance_client/src/app/user/data/types/purchase-order.ts
export interface PurchaseOrder {
  title?: string;
  correspondedInvoiceNumber?: number;
  curentVatAmount?: number;
  curentVatWitholding?: number;
  currentIncomeWithholdingamount?: number;
  currentPoTotalAmount?: number;
  currentTotalAfterTax?: number;
  currentTotalBeforeTax?: number;
  currentVatAmount?: number;
  deletedBy?: string;
  deletedFlag?: string;
  deletedTime?: Date;
  expenseId?: number;
  id?: number;
  inProgress?: boolean;
  incomeWithholdingamount?: number;
  isSent?: boolean;
  modifiablePoParticulars?: [
    {
      amountBalance?: number;
      amountTobepaid?: number;
      deliveryStatus?: string;
      expenseId?: string;
      id?: number;
      incomeWithholdingamount?: number;
      income_w_id?: string;
      invoiceNo?: string;
      incomeTaxRate?: number;
      itemName?: string;
      itemQuantity?: string;
      itemTotalValue?: number;
      itemUnitPrice?: number;
      orderItemType?: string;
      purchaseOrder_id?: string;
      remarks?: string;
      serviceName?: string;
      servicePrice?: string;
      vatAmount?: number;
      vatWitholding?: number;
      vatRate?: number;
      vatWitholdingRate?: number;
      vat_id?: string;
      vat_w_id?: string;
    }
  ];
  modifiedBy?: string;
  modifiedFlag?: string;
  modifiedTime?: Date;
  originalCurrentVatAmount?: number;
  originalIncomeWithholdingamount?: number;
  originalTotalAfterTax?: number;
  originalTotalAmount?: number;
  originalTotalBeforeTax?: number;
  originalVatAmount?: number;
  originalVatWitholding?: number;
  paymentStatus?: true;
  poName?: number;
  poNumber?: string;
  poParticulars?: [
    {
      amountBalance?: number;
      amountTobepaid?: number;
      deliveryStatus?: string;
      expenseId?: string;
      id?: number;
      incomeWithholdingamount?: number;
      incomeTaxRate?: number;
      income_w_id?: string;
      invoiceNo?: string;
      itemName?: string;
      itemQuantity?: string;
      itemTotalValue?: number;
      itemUnitPrice?: number;
      orderItemType?: string;
      purchaseOrder_id?: string;
      remarks?: string;
      serviceName?: string;
      servicePrice?: string;
      vatAmount?: number;
      vatWitholding?: number;
      vatRate?: number;
      vatWitholdingRate?: number;
      vat_id?: string;
      vat_w_id?: string;
    }
  ];
  poStatus?: string;
  poTotalAmount?: number;
  postedBy?: string;
  postedFlag?: string;
  postedTime?: Date;
  reason?: string;
  supplierAddress?: string;
  supplierContact?: string;
  supplierId?: number;
  supplierMobile?: string;
  supplierName?: string;
  totalAfterTax?: number;
  totalBeforeTax?: number;
  vatAmount?: number;
  vatWitholding?: number;
  verifiedBy?: string;
  verifiedFlag?: string;
  verifiedTime?: Date;
}